tags/requirements.txt
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
cssselect==1.2.0
cssselect2==0.7.0
Django==5.0.3
djangorestframework==3.15.1
fake-useragent==1.5.1
html5lib==1.1
idna==3.6
importlib_metadata==7.1.0
lxml==5.2.0
lxml_html_clean==0.1.0
oscrypto==1.3.0
parse==1.20.1
pillow==10.3.0
pycparser==2.22
pyee==11.1.0
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
pypdf==4.1.0
pypng==0.20220715.0
pyppeteer==2.0.0
pyquery==2.0.0
python-bidi==0.4.2
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==1.26.18
w3lib==2.1.2
webencodings==0.5.1
websockets==10.4
xhtml2pdf==0.2.15
zipp==3.18.1
